add new component TableNavigation
